(fix) Fixes datatypes for function in header for local mapper

+++ src/kompass_cpp/kompass_cpp/include/mapping/local_mapper.h
@@ -37,10 +37,10 @@ Eigen::Vector2i localToGrid(const Eigen::Vector2f &poseTargetInCentral,
  * @return Transformed grid.
  */
 Eigen::MatrixXf getPreviousGridInCurrentPose(
-    const Eigen::Vector2d &current_position_in_previous_pose,
+    const Eigen::Vector2f &current_position_in_previous_pose,
     double current_orientation_in_previous_pose,
     const Eigen::MatrixXf &previous_grid_data,
-    const Eigen::Vector2d &central_point, int grid_width, int grid_height,
+    const Eigen::Vector2i &central_point, int grid_width, int grid_height,
     float resolution, float unknown_value);
 
 /**
